requirements.txt
python-telegram-bot==20.8
aiohttp==3.9.3
SQLAlchemy==2.0.29
aiosqlite==0.19.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-dotenv==1.0.1
APScheduler==3.10.4 